* 'master' of https://github.com/guhuohua/pig:
  修改购物车，订单接口，增加地址管理接口

+++ src/main/java/com/ch/entity/GoodsSolrSchema.java
@@ -27,10 +27,10 @@ public class GoodsSolrSchema implements Serializable {
     private String name;
 
     @Field
-    private BigDecimal originalPrice;
+    private Long originalPrice;
 
     @Field
-    private BigDecimal presentPrice;
+    private Long presentPrice;
 
     @Field
     private Integer salesVolume;
